+++ flutter_plugin_generator/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Added support to filter which platforms the plugin has support.
 - Renamed build.dart to flutter_plugin_generator.dart to conform with pub.dev warning.
-
+- **Breaking change** Now the generated plugin will have a static const MethodChannel shared across instances, unless
+at least one path replacement is the in the [MethodCallPlugin.channelName] string.
 ## [0.0.4]
 
 - Removed Flutter dependency.